remove dashboards list from useDashboard
