+++ packages/tupaia-web/src/features/Dashboard/DashboardMenu/DashboardMenu.tsx
@@ -9,6 +9,7 @@ import { ButtonBase, Menu, MenuItem, Box, Paper } from '@material-ui/core';
 import styled from 'styled-components';
 import { Dashboard } from '../../../types';
 import { TOP_BAR_HEIGHT } from '../../../constants';
+import { useDashboards } from '../../../api/queries';
 import { useDashboard } from '../utils';
 import { ActionsMenu } from './ActionsMenu';
 
@@ -80,7 +81,9 @@ const DashboardMenuItem = ({ dashboardName, onClose }: DashboardMenuItemProps) =
 };
 
 export const DashboardMenu = () => {
-  const { dashboards, activeDashboard } = useDashboard();
+  const { projectCode, entityCode } = useParams();
+  const { activeDashboard } = useDashboard();
+  const { data: dashboards = [] } = useDashboards(projectCode, entityCode);
   const [anchorEl, setAnchorEl] = useState<HTMLElement | null>(null);
 
   const handleClickListItem = (event: React.MouseEvent<HTMLElement>) => {
