+++ packages/tupaia-web/src/features/Dashboard/utils/DashboardContext.tsx
@@ -2,10 +2,7 @@
  * Tupaia
  * Copyright (c) 2017 - 2023 Beyond Essential Systems Pty Ltd
  */
-import React, { Dispatch, createContext, useContext, useReducer } from 'react';
-import { useParams } from 'react-router';
-import { useDashboards } from '../../../api/queries';
-import { useGAEffect } from '../../../utils';
+import React, { Dispatch, createContext, useReducer } from 'react';
 
 type DashboardReducer = {
   selectedDashboardItems: string[];
@@ -19,9 +16,9 @@ const defaultState = {
   subscribeModalOpen: false,
 } as DashboardReducer;
 
-const DashboardContext = createContext(defaultState);
+export const DashboardContext = createContext(defaultState);
 
-enum Actions {
+export enum Actions {
   SET_SELECTED_DASHBOARD_ITEMS = 'SET_SELECTED_DASHBOARD_ITEMS',
   TOGGLE_EXPORT_MODAL = 'TOGGLE_EXPORT_MODAL',
   TOGGLE_SUBSCRIBE_MODAL = 'TOGGLE_SUBSCRIBE_MODAL',
@@ -31,7 +28,7 @@ interface ActionT {
   type: Actions;
   payload?: string[];
 }
-const DashboardDispatchContext = createContext<Dispatch<ActionT> | null>(null);
+export const DashboardDispatchContext = createContext<Dispatch<ActionT> | null>(null);
 
 const dashboardReducer = (state: DashboardReducer, action: ActionT): DashboardReducer => {
   switch (action.type) {
@@ -53,45 +50,6 @@ const dashboardReducer = (state: DashboardReducer, action: ActionT): DashboardRe
     default:
       return state;
   }
-};
-
-// Utility hook to get the dashboard context
-// Contains the dashboards, active dashboard, and export and subscribe state
-export const useDashboard = () => {
-  const { projectCode, entityCode, dashboardName } = useParams();
-  const { data: dashboards = [], ...dashboardResult } = useDashboards(projectCode, entityCode);
-  const { selectedDashboardItems, exportModalOpen, subscribeModalOpen } =
-    useContext(DashboardContext);
-  const dispatch = useContext(DashboardDispatchContext);
-
-  const setSelectedDashboardItems = (items: string[]) => {
-    dispatch?.({ type: Actions.SET_SELECTED_DASHBOARD_ITEMS, payload: items });
-  };
-
-  const toggleExportModal = () => {
-    dispatch?.({ type: Actions.TOGGLE_EXPORT_MODAL });
-  };
-
-  const toggleSubscribeModal = () => {
-    dispatch?.({ type: Actions.TOGGLE_SUBSCRIBE_MODAL });
-  };
-  // trim dashboard name to avoid issues with trailing or leading spaces
-  const activeDashboard =
-    dashboards?.find(dashboard => dashboard.name.trim() === dashboardName?.trim()) ?? undefined;
-
-  useGAEffect('Dashboard', 'Change Tab', activeDashboard?.name);
-
-  return {
-    ...dashboardResult,
-    dashboards,
-    activeDashboard,
-    selectedDashboardItems,
-    setSelectedDashboardItems,
-    toggleExportModal,
-    exportModalOpen,
-    subscribeModalOpen,
-    toggleSubscribeModal,
-  };
 };
 
 // A wrapper containing the context providers for the dashboard
