+++ packages/tupaia-web/src/features/Dashboard/utils/index.ts
@@ -3,5 +3,6 @@
  *  Copyright (c) 2017 - 2024 Beyond Essential Systems Pty Ltd
  */
 
-export * from './DashboardContext';
+export { DashboardContextProvider } from './DashboardContext';
 export { useDashboardMailingList } from './useDashboardMailingList';
+export { useDashboard } from './useDashboard';
